Reorginize code and add parametrs
